opmaak aangepast en afbeeldingen uit database halen toegevoegd
