+++ src/components/navigation/Nav.css
@@ -8,7 +8,6 @@
 
 nav ul a,
 nav ul a:visited {
-
     text-decoration: none;
 }
 
@@ -20,97 +19,81 @@ nav a {
 
 nav ul {
     display: flex;
-    padding-right: 40px;
-    padding-bottom: 5px;
     justify-content: space-between;
+    align-items: center; /* Align items vertically in the middle */
     list-style-type: none;
     border-bottom: #242C2E solid 1px;
-    /*border: 5px solid red;*/
+    padding: 0 40px; /* Adjust padding */
 }
-
 
 .searchbar {
     border-radius: 9px;
     width: 600px;
     height: 40px;
-    /*background-color: hsla(200, 26%, 35%, 0.8);*/
     border: none;
     background-color: #1A282D;
     color: white;
-
 }
 
 .mainlogo {
     display: flex;
     align-items: center;
-    justify-content: space-around;
-    /*border: 3px solid white;*/
 }
 
 .brandname {
     display: flex;
     align-items: center;
-    /*border: 3px solid white;*/
 }
 
 .search {
     display: flex;
     align-items: center;
-    justify-self: center;
-    /*border: 3px solid white;*/
-
 }
 
 .navlogin {
-    /*display: flex;*/
-    align-items: center;
     position: relative;
     display: inline-block;
-    /*border: 3px solid white;*/
 }
 
 .navprofile {
     display: flex;
     align-items: center;
-    /*border: 3px solid white;*/
 }
 
 .dropbtn {
-background: none;
-    /*padding: 14px;*/
+    background: none;
     border: none;
-    /*transition: background-color 0.3s ease, color 0.3s ease;*/
     box-shadow: none;
 }
 
 .dropdown-content2 {
     display: none;
     position: absolute;
-
     z-index: 1;
     transition: opacity 0.7s ease;
 }
 
 .dropdown-content2 a {
- font-size: 30px;
+    font-size: 30px;
 }
 
-
 /* Change color of dropdown links on hover */
-.dropdown-content2 a:hover {border-bottom: #242C2E solid 1px;}
+.dropdown-content2 a:hover {
+    border-bottom: #242C2E solid 1px;
+}
 
 /* Show the dropdown menu on hover */
-.navlogin:hover .dropdown-content2 {display: block;}
+.navlogin:hover .dropdown-content2 {
+    display: block;
+}
 
 /* Change the background color of the dropdown button when the dropdown content is shown */
-/*.dropdown:hover .dropbtn {background-color: var(--onHoverColor)}*/
-.navlogin:hover .dropbtn {color:black}
-/*------Drop down menu end------*/
-
+.navlogin:hover .dropbtn {
+    color: black;
+}
 
 @media screen and (max-width: 600px) {
     nav ul {
-        /*display: none;*/
         flex-direction: column;
         padding: 0;
     }
@@ -118,13 +101,13 @@ background: none;
     nav li {
         width: 100%;
         text-align: center;
-
     }
 
     .searchbar {
         width: 400px;
         display: flex;
-        /*justify-content: center;*/
+        padding-left: 15px;
+        padding-right: 15px;
     }
 
     .search {
@@ -132,36 +115,159 @@ background: none;
         justify-content: center;
     }
 
-    .dropdown {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        background-color: #242C2E;
-        padding: 10px;
-        position: fixed;
-        top: 0;
-        width: 100%;
-        z-index: 9999; /* Ensure it's above other content */
-    }
 
-    .dropdown-content {
+
+    .dropdown-content2 {
         display: none;
-        text-align: center;
-        background-color: #242C2E;
-        width: 100%;
+        /*position: absolute;*/
+        z-index: 1;
+        transition: opacity 0.7s ease;
+        top: 100%; /* Position below dropdown button */
+        left: -25%; /* Align with left edge of dropdown button */
+        /*background-color: #fff; !* Adjust as needed *!*/
+        border: 1px solid #ccc; /* Adjust as needed */
+        border-radius: 8px;
+        padding: 10px; /* Adjust as needed */
+       ;
     }
 
-    .dropdown:hover .dropdown-content {
+    .dropdown-content2 a {
         display: block;
-    }
-
-    .dropdown-content a {
-        display: block;
-        padding: 10px;
-        color: white;
+        font-size: 30px;
+        color: white; /* Adjust as needed */
         text-decoration: none;
+
+
+    }
+
+    .navlogin:hover .dropdown-content2 {
+        display: block;
+    }
+
+    /* Apply margin to items below dropdown content */
+    .navlogin:hover ~ .navprofile,
+    .navlogin:hover ~ .searchbar {
+        margin-top: 100px; /* Adjust as needed */
     }
 }
+
+
+
+
+
+
+
+/*.default-link h1 {*/
+/*    border-bottom: 1px solid transparent;*/
+/*}*/
+
+/*.active-link h1 {*/
+/*    border-bottom: #242C2E solid 1px;*/
+/*}*/
+
+/*nav ul a,*/
+/*nav ul a:visited {*/
+
+/*    text-decoration: none;*/
+/*}*/
+
+/*nav a {*/
+/*    font-size: 35px;*/
+/*    color: white;*/
+/*    font-weight: 700;*/
+/*}*/
+
+/*nav ul {*/
+/*    display: flex;*/
+/*    padding-right: 40px;*/
+/*    padding-left: 40px;*/
+/*    padding-bottom: 5px;*/
+/*    justify-content: space-between;*/
+/*    list-style-type: none;*/
+/*    border-bottom: #242C2E solid 1px;*/
+/*    !*border: 5px solid red;*!*/
+/*}*/
+
+
+/*.searchbar {*/
+/*    border-radius: 9px;*/
+/*    width: 600px;*/
+/*    height: 40px;*/
+/*    !*background-color: hsla(200, 26%, 35%, 0.8);*!*/
+/*    border: none;*/
+/*    background-color: #1A282D;*/
+/*    color: white;*/
+
+/*}*/
+
+/*.mainlogo {*/
+/*    display: flex;*/
+/*    align-items: center;*/
+/*    justify-content: space-around;*/
+/*    !*border: 3px solid white;*!*/
+/*}*/
+
+/*.brandname {*/
+/*    display: flex;*/
+/*    align-items: center;*/
+/*    !*border: 3px solid white;*!*/
+/*}*/
+
+/*.search {*/
+/*    display: flex;*/
+/*    align-items: center;*/
+/*    justify-self: center;*/
+/*    !*border: 3px solid white;*!*/
+
+/*}*/
+
+/*.navlogin {*/
+/*    !*display: flex;*!*/
+/*    align-items: center;*/
+/*    position: relative;*/
+/*    display: inline-block;*/
+/*    !*border: 3px solid white;*!*/
+/*}*/
+
+/*.navprofile {*/
+/*    display: flex;*/
+/*    align-items: center;*/
+/*    !*border: 3px solid white;*!*/
+/*}*/
+
+/*.dropbtn {*/
+/*background: none;*/
+/*    !*padding: 14px;*!*/
+/*    border: none;*/
+/*    !*transition: background-color 0.3s ease, color 0.3s ease;*!*/
+/*    box-shadow: none;*/
+/*}*/
+
+/*.dropdown-content2 {*/
+/*    display: none;*/
+/*    position: absolute;*/
+
+/*    z-index: 1;*/
+/*    transition: opacity 0.7s ease;*/
+/*}*/
+
+/*.dropdown-content2 a {*/
+/* font-size: 30px;*/
+/*}*/
+
+
+/*!* Change color of dropdown links on hover *!*/
+/*.dropdown-content2 a:hover {border-bottom: #242C2E solid 1px;}*/
+
+/*!* Show the dropdown menu on hover *!*/
+/*.navlogin:hover .dropdown-content2 {display: block;}*/
+
+/*!* Change the background color of the dropdown button when the dropdown content is shown *!*/
+/*!*.dropdown:hover .dropbtn {background-color: var(--onHoverColor)}*!*/
+/*.navlogin:hover .dropbtn {color:black}*/
+/*!*------Drop down menu end------*!*/
+
+
 
 /*@media screen and (max-width: 600px) {*/
 /*    nav ul {*/
