+++ src/components/post/Post.css
@@ -1,5 +1,3 @@
-
-
 .postContainerOuter {
     display: flex;
     justify-content: center;
@@ -10,9 +8,9 @@
 }
 
 
-
 .inner-content-container {
     width: 80%;
+    border-radius: 8px;
 }
 
 .blog-post {
@@ -20,6 +18,7 @@
     padding: 20px;
     margin: 20px 0;
     box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    border-radius: 8px;
 }
 
 .blog-post h2 {
@@ -45,4 +44,59 @@
     display: inline-block;
     margin-top: 10px;
 }
+
+
+@media screen and (max-width: 768px) {
+    .postContainerOuter {
+        align-items: center; /* Adjust alignment for mobile */
+        padding: 20px; /* Add padding to give some space around the content */
+    }
+
+    .inner-content-container {
+
+        width: 100%; /* Adjust width for mobile */
+    }
+}
+
+/*@media screen and (max-width: 768px) {*/
+/*    .postContainerOuterSolo {*/
+/*        align-items: center;*/
+/*    }*/
+
+/*    .inner-content-container {*/
+/*        width: 90%;*/
+/*        max-width: 800px;*/
+/*        display: flex;*/
+/*        flex-direction: column;*/
+/*        align-items: center;*/
+/*    }*/
+
+/*    .blog-post {*/
+/*        width: 100%;*/
+/*        padding: 10px;*/
+/*        margin-bottom: 20px;*/
+/*    }*/
+
+/*    .blog-post h2 {*/
+/*        font-size: 20px;*/
+/*        margin: 0;*/
+/*    }*/
+
+/*    .blog-post p {*/
+/*        font-size: 14px;*/
+/*        margin: 0;*/
+/*    }*/
+
+/*    .blog-post img {*/
+/*        max-width: 100%;*/
+/*        border-radius: 8px;*/
+/*    }*/
+
+/*    .profilelink  {*/
+/*        color: #0079d3;*/
+
+
+/*    }*/
+
+
