+++ src/index.css
@@ -1,6 +1,11 @@
 @import url('https://fonts.googleapis.com/css2?family=Geologica:wght@300&family=Roboto&display=swap');
 @import url('https://fonts.googleapis.com/css?family=Open+Sans:300');
 
+* {
+    padding: 0;
+    margin: 0;
+    box-sizing: border-box;
+}
 :root {
 
 }
