+++ src/pages/makePost/MakePost.css
@@ -13,6 +13,23 @@
     max-width: 600px; /* Optioneel: voeg een maximale breedte toe aan de container */
     width: 100%; /* Zorg ervoor dat de container de volledige breedte van de beschikbare ruimte inneemt */
     padding: 20px;
+    border-radius: 8px;
+}
+
+.postbutton {
+
+        border: 1px solid rgba(10, 180, 180, 1);
+        background: rgba(20, 20, 20, .6);
+        font-size: 18px;
+        color: white;
+        margin-top: 20px;
+        padding: 10px 50px;
+        cursor: pointer;
+        transition: .4s;
+        &:hover {
+            background: rgba(20, 20, 20, .8);
+            padding: 10px 80px;
+        }
 }
