+++ src/pages/profile/Profile.css
@@ -17,3 +17,19 @@
     margin: 10px 0;
     color: #666;
 }
+
+.profile-container img {
+    width: 100px;
+    height: 100px;
+    border-radius: 50%;
+    margin-bottom: 20px;
+}
+
+.fundiv {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    height: 10vh;
+
+}
