+++ src/pages/profile/Profile.jsx
@@ -3,10 +3,81 @@ import {AuthContext} from "../../context/AuthContextProvider.jsx";
 import {useContext, useEffect, useState} from "react";
 import axios from 'axios';
 import {Link} from "react-router-dom";
+import useProfileImage from "../../hooks/useProfileImage.jsx";
 
 function Profile() {
     const [profileData, setProfileData] = useState({});
-    const { user } = useContext(AuthContext);
+    const {user} = useContext(AuthContext);
+    const [selectedFile, setSelectedFile] = useState({});
+    const [warning, setWarning] = useState('');
+    // const [download, triggerDownload] = useState(false);
+    const {profileImage} = useProfileImage();
+
+
+    const handleFileChange = (event) => {
+        const selected = event.target.files[0];
+        setWarning('');
+        // Controleer of er een bestand is geselecteerd
+        if (selected) {
+            // Gebruik een FileReader om het bestand in te lezen en de afmetingen te controleren
+            const reader = new FileReader();
+            reader.readAsDataURL(selected);
+            reader.onload = (e) => {
+                const img = new Image();
+                img.src = e.target.result;
+                img.onload = () => {
+                    // Afbeelding limiet in pixels
+                    if (img.width <= 1000 && img.height <= 1000) {
+                        // Controleer of het bestand de maximale grootte niet overschrijdt (10000 bytes is gelijk aan 10 kilobytes (KB))
+                        if (selected.size <= 10000000) {
+                            // Als het bestand binnen de grenzen valt, stel het in als geselecteerd bestand
+                            setSelectedFile(selected);
+                            console.log(selected);
+                        } else {
+                            setWarning('Maximale toegestane grootte is 10 KB');
+                        }
+                    } else {
+                        setWarning('Afbeelding is groter dan 1000 x 1000 pixels.');
+                    }
+                };
+            };
+        }
+    };
+
+
+    async function deleteOldImage() {
+        try {
+            axios.delete(`http://localhost:8080/image/${user.username}`);
+        } catch (error) {
+            console.error("Fout bij het verwijderen van de oude afbeelding:", error);
+        }
+    }
+
+    async function uploadImage() {
+        if (selectedFile) {
+            // Verwijder de oude afbeelding
+            deleteOldImage();
+
+            // Upload de nieuwe afbeelding nadat de oude is verwijderd
+            const formData = new FormData();
+            formData.append('file', selectedFile);
+            formData.append('username', user.username);
+
+            try {
+                const response = await axios.post('http://localhost:8080/image', formData, {
+                    headers: {
+                        "Content-Type": "multipart/form-data",
+                    }
+                });
+                console.log(response);
+            } catch (error) {
+                console.error("Fout bij het uploaden van de afbeelding:", error);
+            }
+        } else {
+            console.warn("Selecteer een afbeelding om te uploaden.");
+        }
+    }
+
 
     useEffect(() => {
         // we halen de pagina-content op in de mounting-cycle
@@ -31,10 +102,8 @@ function Profile() {
     }, [])
 
 
-
     return (
         <>
-
 
 
             {/*{Object.keys(profileData).length > 0 &&*/}
@@ -43,22 +112,37 @@ function Profile() {
 
                     <div className="profile-container">
                         <h2>Profile</h2>
-                        <p> <strong>Email:</strong>  {user.email}</p>
-                        <p> <strong>Username:</strong> {user.username}</p>
+                        <p><strong>Email:</strong> {user.email}</p>
+                        <p><strong>Username:</strong> {user.username}</p>
                         <p><strong>Password:</strong> ******** </p>
                         <p><strong>User role:</strong> {user.role} </p>
 
+                        <input
+                            className="ProfilePictureUpload"
+                            type="file"
+                            accept=".jpg, .jpeg, .png"
+                            name="profilePhotoUpload"
+                            placeholder="Upload een profielfoto"
+                            id="profilePhotoUpload"
+                            onChange={handleFileChange}
+                        />
+
+
+                        <div className="profile-image">
+                            {profileImage && <img src={profileImage} alt="Profiel foto"
+                                                  style={{width: '200px', height: '200px'}}/>}
+                        </div>
+                        <button type="submit" onClick={uploadImage}>Upload mij!</button>
                     </div>
 
-
-                    <h2>Strikt geheime profiel-content</h2>
-                    <h3>{profileData.title}</h3>
-                    <p>{profileData.content}</p>
-
+                    <div className="fundiv">
+                        <h2>Only you (and the FBI) can see this information</h2>
+                        <p>Back to the <Link to="/">Homepage</Link></p>
+                    </div>
 
                 </section>
             }
-            <p>Terug naar de <Link to="/">Homepagina</Link></p>
+
 
         </>
     );
