+++ src/pages/profilepost/ProfilePost.css
@@ -1,0 +1,73 @@
+.postContainerOuterSolo {
+    display: flex;
+    justify-content: center;
+    margin-top: 50px;
+}
+
+.inner-content-containerSolo {
+    width: 90%; /* Adjust width as needed */
+    max-width: 800px; /* Adjust max-width as needed */
+}
+
+.blog-post-solo {
+    border: 1px solid #ccc;
+    border-radius: 8px;
+    box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1);
+    padding: 10px;
+    margin-bottom: 20px;
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    gap: 10px; /* Adjust gap between elements */
+}
+
+.blog-post-solo h2 {
+    font-size: 24px;
+    margin: 0; /* Remove margin */
+}
+
+.blog-post-solo p {
+    font-size: 16px;
+    margin: 0; /* Remove margin */
+}
+
+.blog-post-solo img {
+    max-width: 100%;
+    border-radius: 8px;
+}
+
+@media screen and (max-width: 768px) {
+    .postContainerOuterSolo {
+        align-items: center;
+    }
+
+    .inner-content-containerSolo {
+        width: 90%;
+        max-width: 800px;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+    }
+
+    .blog-post-solo {
+        width: 100%;
+        padding: 10px;
+        margin-bottom: 20px;
+    }
+
+    .blog-post-solo h2 {
+        font-size: 20px;
+        margin: 0;
+    }
+
+    .blog-post-solo p {
+        font-size: 14px;
+        margin: 0;
+    }
+
+    .blog-post-solo img {
+        max-width: 100%;
+        border-radius: 8px;
+    }
+}
+
